+++ package.json
@@ -72,10 +72,12 @@
 	},
 	"dependencies": {
 		"@babel/runtime": "^7.21.0",
+		"@fortawesome/fontawesome-free": "^6.4.0",
 		"bootstrap": "^5.2.3",
 		"dotenv": "^16.0.3",
 		"express": "^4.18.2",
 		"helmet": "^6.0.1",
+		"mdb-react-ui-kit": "^6.0.0",
 		"morgan": "^1.10.0",
 		"pg": "^8.10.0",
 		"react-bootstrap": "^2.7.4",
